Split out SetupKubectlAuthentication
