+++ source/Calamari.Tests/KubernetesFixtures/KubernetesRunsScriptFixture.cs
@@ -15,6 +15,7 @@ using Calamari.Common.FeatureToggles;
 using Calamari.Common.Plumbing.FileSystem;
 using Calamari.Common.Plumbing.Variables;
 using Calamari.Kubernetes;
+using Calamari.Kubernetes.Context;
 using Calamari.Testing;
 using Calamari.Testing.Helpers;
 using Calamari.Testing.Requirements;
