+++ source/Calamari/Kubernetes/Context/KubeContextConstants.cs
@@ -1,0 +1,9 @@
+namespace Calamari.Kubernetes.Context
+{
+    public static class KubeContextConstants
+    {
+        public const string User = "octouser";
+        public const string Cluster = "octocluster";
+        public const string Context = "octocontext";
+    }
+}
